+++ netket/experimental/observable/__init__.py
@@ -12,12 +12,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from .runtime_log import RuntimeLog
-from .json_log import JsonLog
-from .state_log import StateLog
-from .tensorboard import TensorBoardLog
-
-
-from netket.utils import _hide_submodules
-
-_hide_submodules(__name__)
+from .renyi2 import Renyi2EntanglementEntropy